% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_states.R
\name{abbreviate_states}
\alias{abbreviate_states}
\title{Abbreviate Australian state names}
\usage{
abbreviate_states(state_names)
}
\arguments{
\item{state_names}{vector of state names in long form}
}
\value{
shortened state names
}
\description{
Abbreviate Australian state names
}
\examples{
abbreviate_states("Victoria")
}
