% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_age_work.R
\name{abs_age_work_lga}
\alias{abs_age_work_lga}
\alias{abs-age-work}
\alias{abs_age_work_state}
\title{Return  data on employed population for a given age and state or
lga of Australia}
\usage{
abs_age_work_lga(lga = NULL, age = NULL)

abs_age_work_state(state = NULL, age = NULL)
}
\arguments{
\item{lga}{target Australian local government area (LGA) name, such as
"Fairfield (C)" or a vector with multiple lga names. See
\code{\link[=abs_lga_lookup]{abs_lga_lookup()}} for list of lga names.}

\item{age}{a numeric or numeric vector denoting ages between 0 to 115.
The default is to return all ages.}

\item{state}{target Australian state name or a vector with multiple state
names in its abbreviated form, such as "QLD", "NSW", or "TAS"}
}
\value{
data set with information on the number of employed people belonging
to a particular age, its total population and the corresponding proportion.
}
\description{
Return  data on employed population for a given age and state or
lga of Australia
}
\examples{
abs_age_work_state(state = "NSW")
abs_age_work_state(state = c("QLD", "TAS"), age = 5)
abs_age_work_lga(lga = "Albany (C)", age = 1:5)
abs_age_work_lga(lga = c("Albury (C)", "Barcoo (S)"), age = 39)

}
