% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_states.R
\name{abs_abbreviate_states}
\alias{abs_abbreviate_states}
\title{Abbreviate Australian State Names}
\usage{
abs_abbreviate_states(state_names)
}
\arguments{
\item{state_names}{vector of state names in long form}
}
\value{
shortened state names
}
\description{
Given a full name (Title Case) of an Australian State or Territory, produces
the abbreviated state name.
}
\examples{
abs_abbreviate_states("Victoria")
abs_abbreviate_states(c("Victoria", "Queensland"))
}
\seealso{
\code{\link[=abs_unabbreviate_states]{abs_unabbreviate_states()}}
}
