% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs-helpers.R
\name{abs_age_data}
\alias{abs_age_data}
\alias{abs_age_lga}
\alias{abs_age_state}
\title{Return Australian Bureau of Statistics (ABS) age population data for a
given Local Government Area (LGA) or state}
\usage{
abs_age_lga(lga_name)

abs_age_state(state_name)
}
\arguments{
\item{lga_name}{lga name - can be a partial match, e.g.,  although the official name might be "Albury (C)", "Albury" is fine. It must also match exactly one LGA. See \code{\link[=check_lga_name]{check_lga_name()}} for more details.}

\item{state_name}{shortened state name}
}
\value{
dataset of: \code{lga} (or \code{state}), \code{lower.age.limit}, \code{year},
and \code{population}.
}
\description{
Return Australian Bureau of Statistics (ABS) age population data for a
given Local Government Area (LGA) or state
}
\examples{
abs_age_lga(c("Albury (C)","Fairfield (C)"))
abs_age_state(c("NSW","VIC"))
}
