% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_age_state.R
\name{abs_age_state}
\alias{abs_age_state}
\title{Return abs age population data for a given state of Australia}
\usage{
abs_age_state(state_name)
}
\arguments{
\item{state_name}{shortened state name}
}
\value{
dataset of columns
}
\description{
Return abs age population data for a given state of Australia
}
\examples{
abs_age_state("NSW")
}
