% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-abs-avg-work-school.R
\docType{data}
\name{abs_avg_school}
\alias{abs_avg_school}
\title{ABS education data for 2016}
\format{
A data frame with 116 rows and 2 variables:
\describe{
\item{age}{0 to 115}
\item{school_fraction}{fraction of population at school}
}
}
\source{
{Census of Population and Housing, 2016, TableBuilder}
}
\usage{
abs_avg_school
}
\description{
An internal dataset containing Australian Bureau of Statistics education data for
each age in 2016. The data is averaged across each state to provide an
overall average, and is used to provide estimated education populations for
model fitting in \code{\link[=add_school_work_participation]{add_school_work_participation()}}, which is used in \code{\link[=fit_single_contact_model]{fit_single_contact_model()}}. The data is summarised from \code{data_abs_state_education},
see \code{?data_abs_state_education} for more details.
}
\keyword{datasets}
