% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-abs-avg-work-school.R
\docType{data}
\name{abs_avg_work}
\alias{abs_avg_work}
\title{ABS work data for 2016}
\format{
A data frame with 116 rows and 2 variables:
\describe{
\item{age}{0 to 115}
\item{work_fraction}{fraction of population working.}
}
}
\source{
{Census of Population and Housing, 2016, TableBuilder}
}
\usage{
abs_avg_work
}
\description{
An internal dataset containing Australian Bureau of Statistics work data for
each age in 2016. The data is averaged across each state to provide an
overall average, and is used to provide estimated work populations for
model fitting in \code{\link[=add_school_work_participation]{add_school_work_participation()}}, which is used in \code{\link[=fit_single_contact_model]{fit_single_contact_model()}}. The data is summarised from \code{data_abs_state_work},
see \code{?data_abs_state_work} for more details.
}
\keyword{datasets}
