% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_household_size_population.R
\name{abs_household_size_population}
\alias{abs_household_size_population}
\title{Get population associated with each household size in an LGA or a state}
\usage{
abs_household_size_population(state = NULL, lga = NULL)
}
\value{
returns a data frame with household size and the population associated with it in each LGA or state.
}
\description{
Get population associated with each household size in an LGA or a state
}
\examples{
abs_household_size_population(state = "NSW")
}
