% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abs_lga_lookup.R
\docType{data}
\name{abs_lga_lookup}
\alias{abs_lga_lookup}
\title{ABS lookup table of states, lga code and lga name}
\format{
A data frame with 557 rows and 3 variables, arrange by state then LGA
\describe{
\item{state}{state - short state or territory name}
\item{lga_code}{official lga code}
\item{lga}{lga name}
}
}
\usage{
abs_lga_lookup
}
\description{
A dataset containing Australian Bureau of Statistics official short state
names, lga_code, and lga name.
}
\keyword{datasets}
