% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_per_capita_household_size_lga.R
\name{abs_per_capita_household_size_lga}
\alias{abs_per_capita_household_size_lga}
\title{Get household size distribution based on  LGA name}
\usage{
abs_per_capita_household_size_lga(lga = NULL)
}
\arguments{
\item{lga}{target Australian local government area (LGA) name, such as "Fairfield (C)".  See
\code{\link[=abs_lga_lookup]{abs_lga_lookup()}} for list of lga names}
}
\value{
returns a data frame with household size distributions of a specific state or LGA.
Data frame contains variables depicting the lga wise household size (number of people in a household) and the associated population.
}
\description{
Get household size distribution based on  LGA name
}
\examples{
abs_per_capita_household_size_lga(lga = c("Fairfield (C)","Albury (C)"))

}
