% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_per_capita_household_size_state.R
\name{abs_per_capita_household_size_state}
\alias{abs_per_capita_household_size_state}
\title{Get household size distribution based on state name}
\usage{
abs_per_capita_household_size_state(state = NULL)
}
\arguments{
\item{state}{target Australian state name in abbreviated form, such as "QLD", "NSW", or "TAS"}
}
\value{
returns a data frame with household size distributions of a specific state.
Data frame contains variables depicting the lga wise household size (number of people in a household) and the associated population.
}
\description{
Get household size distribution based on state name
}
\examples{
abs_per_capita_household_size_state(state = c("NSW","TAS"))
}
