% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_states.R
\name{abs_unabbreviate_states}
\alias{abs_unabbreviate_states}
\title{Un-abbreviate Australian state names}
\usage{
abs_unabbreviate_states(state_names)
}
\arguments{
\item{state_names}{vector of state names in short form}
}
\value{
Longer state names
}
\description{
Un-abbreviate Australian state names
}
\examples{
abs_unabbreviate_states("VIC")
abs_unabbreviate_states(c("VIC", "QLD"))
}
\seealso{
\code{\link[=abs_abbreviate_states]{abs_abbreviate_states()}}
}
