% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmat-population.R
\name{age}
\alias{age}
\alias{age_label}
\alias{age_label.default}
\alias{age_label.conmat_population}
\alias{population_label}
\alias{population_label.default}
\alias{population_label.conmat_population}
\alias{population}
\title{Accessing conmat attributes}
\usage{
age(x)

age_label(x)

\method{age_label}{default}(x)

\method{age_label}{conmat_population}(x)

population_label(x)

\method{population_label}{default}(x)

\method{population_label}{conmat_population}(x)

population(x)
}
\arguments{
\item{x}{conmat_population data frame}
}
\value{
age or population symbol or label
}
\description{
Accessing conmat attributes
}
\examples{
perth <- abs_age_lga("Perth (C)")
age(perth)
age_label(perth)
population(perth)
population_label(perth)
}
