% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modelling_features.R
\name{add_intergenerational}
\alias{add_intergenerational}
\title{Add column, "intergenerational"}
\usage{
add_intergenerational(data)
}
\arguments{
\item{data}{data.frame with columns \code{age_from}, and \code{age_to}}
}
\value{
data.frame with extra column, \code{intergenerational}
}
\description{
For modelling purposes it is useful to have a feature that is the absolute
difference between \code{age_from} and \code{age_to} columns.
}
\examples{

polymod_contact <- get_polymod_contact_data()

polymod_contact \%>\% add_intergenerational()

}
