% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modelling_features.R
\name{add_modelling_features}
\alias{add_modelling_features}
\title{FUNCTION_TITLE}
\usage{
add_modelling_features(contact_data, ...)
}
\arguments{
\item{contact_data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
