% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_offset.R
\name{add_offset}
\alias{add_offset}
\title{Adds offset}
\usage{
add_offset(contact_data)
}
\arguments{
\item{contact_data}{contact_data}
}
\value{
contact data with two extra columns:
\code{log_contactable_population_school} and \code{log_contactable_population}
}
\description{
Adds offset
}
\author{
Nick Golding
}
