% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_population_age_to.R
\name{add_population_age_to}
\alias{add_population_age_to}
\title{FUNCTION_TITLE}
\usage{
add_population_age_to(contact_data, population = get_polymod_population())
}
\arguments{
\item{contact_data}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION, Default: get_polymod_population()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
