% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rotated_ages.R
\name{add_rotated_ages}
\alias{add_rotated_ages}
\title{FUNCTION_TITLE}
\usage{
add_rotated_ages(contact_data)
}
\arguments{
\item{contact_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
