% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_school_work_participation.R
\name{add_school_work_participation}
\alias{add_school_work_participation}
\title{FUNCTION_TITLE}
\usage{
add_school_work_participation(contact_data)
}
\arguments{
\item{contact_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{across}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{case_when}}
}
