% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symmetrical_features.R
\name{add_symmetrical_features}
\alias{add_symmetrical_features}
\title{Add symmetrical, age based features}
\usage{
add_symmetrical_features(data)
}
\arguments{
\item{data}{data.frame with columns, \code{age_from}, and \code{age_to}}
}
\value{
data.frame with 6 more columns, \code{gam_age_offdiag}, \code{gam_age_offdiag_2}, \code{gam_age_diag_prod}, \code{gam_age_diag_sum}, \code{gam_age_pmax}, \code{gam_age_pmin},
}
\description{
This function adds 6 columns to assist with describing
various age based interactions for model fitting. Requires that the
age columns are called "age_from", and "age_to"
}
\examples{
vec_age <- 0:2
dat_age <- expand.grid(
  age_from = vec_age,
  age_to = vec_age
)

add_symmetrical_features(dat_age)

}
