% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_household_contact_matrix.R
\name{adjust_household_contact_matrix}
\alias{adjust_household_contact_matrix}
\title{Adjust Household Contact Matrix}
\usage{
adjust_household_contact_matrix(
  setting_matrices,
  population,
  household_size,
  household_contact_rate = 1,
  model_mean_other_householders = 2.248971
)
}
\arguments{
\item{setting_matrices}{setting matrix}

\item{population}{population information}

\item{household_size}{single number of the size of a household}

\item{household_contact_rate}{default of 1}

\item{model_mean_other_householders}{default of 2.248971}
}
\value{
contact matrix that has household adjustment
}
\description{
Adjust Household Contact Matrix
}
\author{
Nick Golding
}
