% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{age_breaks}
\alias{age_breaks}
\alias{age_breaks.conmat_age_matrix}
\alias{age_breaks.default}
\title{Extract age break attribute information}
\usage{
age_breaks(matrix)

\method{age_breaks}{conmat_age_matrix}(matrix)

\method{age_breaks}{default}(matrix)
}
\arguments{
\item{matrix}{a \code{conmat_age_matrix} matrix}
}
\value{
age breaks character vector
}
\description{
Extract age break attribute information
}
\section{Methods (by class)}{
\itemize{
\item \code{age_breaks(conmat_age_matrix)}: Get age break information

\item \code{age_breaks(default)}: Get age break information

}}
\examples{
#' age_break_names <- c("[0,5)", "[5,10)", "[10, 15)")
age_mat <- matrix(
  runif(9),
  nrow = 3,
  ncol = 3,
  dimnames = list(
    age_break_names,
    age_break_names
  )
)

age_mat <- new_age_matrix(age_mat)

age_breaks(age_mat)
}
