% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_age_group_lookup.R
\docType{data}
\name{age_group_lookup}
\alias{age_group_lookup}
\title{Lookup table of age groups in 5 year bins}
\format{
A data frame with 21 rows and 3 variables:
\describe{
\item{lower}{Lower age}
\item{upper}{upper age}
\item{age_group}{age group as a factor}
}
}
\usage{
age_group_lookup
}
\description{
A dataset containing age lower and upper levels with age group
}
\keyword{datasets}
