% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-population-year.R
\name{age_population}
\alias{age_population}
\title{Return cleaned population data with lower and upper limits of age}
\usage{
age_population(
  data,
  country_col = NULL,
  country = NULL,
  age_col,
  year_col = NULL,
  year = NULL
)
}
\arguments{
\item{data}{dataset}

\item{country_col}{column name for country}

\item{country}{country name}

\item{age_col}{column name for age}

\item{year_col}{column name for year}

\item{year}{year}
}
\value{
dataset of columns
}
\description{
Return cleaned population data with lower and upper limits of age
}
\examples{
\dontrun{ 
age_population(
  data = data,
  country_col = country,
  country = "Australia",
  age_col = age,
  year_col = year,
  year = 2015
)
age_population(
  data = data,
  country_col = country,
  country = "Australia",
  age_col = age
)
  age_population(
  data = data,
  age_col = age,
  year_col = year,
  year = 2009
)

}
}
