% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-population-year.R
\name{age_population_year}
\alias{age_population_year}
\title{Return cleaned population data with lower and upper limits of age}
\usage{
age_population_year(data, age_col, year_col, year)
}
\arguments{
\item{data}{dataset}

\item{age_col}{column name for age}

\item{year_col}{column name for year}

\item{year}{year}
}
\value{
dataset of columns
}
\description{
Return cleaned population data with lower and upper limits of age
}
\examples{
age_population_year(
  data = worldwide_data,
  age_col = age,
  year_col = year,
  year = 2015
)
}
