% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_population_year_country.R
\name{age_population_year_country}
\alias{age_population_year_country}
\title{Return cleaned country specific population data with lower and upper limits of age}
\usage{
age_population_year_country(
  data,
  country_col,
  country,
  age_col,
  year_col,
  year
)
}
\arguments{
\item{data}{dataset}

\item{country_col}{column name for country}

\item{country}{country name}

\item{age_col}{column name for age}

\item{year_col}{column name for year}

\item{year}{year}
}
\value{
dataset of columns
}
\description{
Return cleaned country specific population data with lower and upper limits of age
}
\examples{
age_population_year_country(
  data = worldwide_data,
  country_col = country,
  country = "Australia",
  age_col = age,
  year_col = year,
  year = 2015
)
}
