% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_predicted_contacts.R
\name{aggregate_predicted_contacts}
\alias{aggregate_predicted_contacts}
\title{FUNCTION_TITLE}
\usage{
aggregate_predicted_contacts(
  predicted_contacts_1y,
  population,
  age_breaks = c(seq(0, 75, by = 5), Inf)
)
}
\arguments{
\item{predicted_contacts_1y}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}

\item{age_breaks}{PARAM_DESCRIPTION, Default: c(seq(0, 75, by = 5), Inf)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
\code{\link[stats]{weighted.mean}}
}
