% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-prediction-helpers.R
\name{autoplot.partial_predictions_sum}
\alias{autoplot.partial_predictions_sum}
\alias{autoplot-conmat-partial}
\alias{autoplot.partial_predictions}
\alias{autoplot.setting_partial_predictions}
\title{Plot partial predictive plots using ggplot2}
\usage{
\method{autoplot}{partial_predictions_sum}(object, ...)

\method{autoplot}{partial_predictions}(object, ...)

\method{autoplot}{setting_partial_predictions}(object, ...)
}
\arguments{
\item{object}{An object with partial predictions from}

\item{...}{Other arguments passed on}
}
\value{
a ggplot visualisation of partial effects
}
\description{
Plot partial predictive plots using ggplot2
}
