% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot-conmat}
\alias{autoplot-conmat}
\alias{autoplot.conmat_age_matrix}
\alias{autoplot.conmat_setting_prediction_matrix}
\alias{autoplot.transmission_probability_matrix}
\alias{autoplot.ngm_setting_matrix}
\alias{autoplot.setting_vaccination_matrix}
\title{Plot setting matrices using ggplot2}
\usage{
\method{autoplot}{conmat_age_matrix}(object, ..., title = "Contact Matrices")

\method{autoplot}{conmat_setting_prediction_matrix}(object, ..., title = "Setting-specific synthetic contact matrices")

\method{autoplot}{transmission_probability_matrix}(
  object,
  ...,
  title = "Setting-specific transmission probability matrices"
)

\method{autoplot}{ngm_setting_matrix}(object, ..., title = "Setting-specific NGM matrices")

\method{autoplot}{setting_vaccination_matrix}(object, ..., title = "Setting-specific vaccination matrices")
}
\arguments{
\item{object}{A matrix or a list of square matrices, with row and column
names indicating the age groups.}

\item{...}{Other arguments passed on}

\item{title}{Title to give to plot setting matrices. Defaults are provided
for certain objects}
}
\value{
a ggplot visualisation of contact rates
}
\description{
Plot setting matrices using ggplot2
}
\examples{
\dontrun{
if (interactive()) {
  polymod_contact_data <- get_polymod_setting_data()
  polymod_survey_data <- get_polymod_population()

  setting_models <- fit_setting_contacts(
    contact_data_list = polymod_contact_data,
    population = polymod_survey_data
  )

  fairfield <- abs_age_lga("Fairfield (C)")

  fairfield_hh_size <-
    get_abs_per_capita_household_size(lga = "Fairfield (C)")

  synthetic_settings_5y_fairfield_hh <- predict_setting_contacts(
    population = fairfield,
    contact_model = setting_models,
    age_breaks = c(seq(0, 85, by = 5), Inf),
    per_capita_household_size = fairfield_hh_size
  )

  # Plotting synthetic contact matrices across all settings

  autoplot(
    object = synthetic_settings_5y_fairfield_hh,
    title = "Setting specific synthetic contact matrices"
  )

  # Work setting specific synthetic contact matrices
  autoplot(
    object = synthetic_settings_5y_fairfield_hh$work,
    title = "Work"
  )
}
}
}
