% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_widths.R
\name{bin_widths}
\alias{bin_widths}
\title{return the widths of bins denoted by a sequence of lower bounds}
\usage{
bin_widths(lower_bound)
}
\arguments{
\item{lower_bound}{lower bound value - a numeric vector}
}
\value{
vector
}
\description{
return the widths of bins denoted by a sequence of lower
bounds (assumming the final is the same as the \verb{[penultimate]}).
}
\author{
Nick Golding
}
