% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_dimensions}
\alias{check_dimensions}
\title{Check dimensions}
\usage{
check_dimensions(ngm, data)
}
\arguments{
\item{ngm}{list with next generation matrices at different settings}

\item{data}{data frame}
}
\description{
An internal function used within \code{\link[=apply_vaccination]{apply_vaccination()}} to warn users of incompatible dimensions of
data and the next generation matrices
}
\keyword{internal}
