% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_if_list}
\alias{check_if_list}
\title{Check if data is a list}
\usage{
check_if_list(contact_data)
}
\arguments{
\item{contact_data}{data on the contacts between two ages at different settings}
}
\description{
Check if data is a list
}
\keyword{internal}
