% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_work_school_demographics.R
\name{check_school_demographics}
\alias{check_school_demographics}
\title{Check School Demographics}
\usage{
check_school_demographics(school_demographics)
}
\arguments{
\item{school_demographics}{school data}
}
\description{
Check School Demographics
}
\author{
njtierney
}
\keyword{internal}
