% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_state_name.R
\name{check_state_name}
\alias{check_state_name}
\title{Check state name in Australia}
\usage{
check_state_name(state_name)
}
\arguments{
\item{state_name}{character of length 1}
}
\value{
errors if state name not in ABS data
}
\description{
Check state name in Australia
}
