% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_work_school_demographics.R
\name{check_work_demographics}
\alias{check_work_demographics}
\title{Check Work Demographics}
\usage{
check_work_demographics(work_demographics)
}
\arguments{
\item{work_demographics}{work data}
}
\description{
Check Work Demographics
}
\author{
njtierney
}
\keyword{internal}
