% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_age_population_year}
\alias{clean_age_population_year}
\title{Prepare age population data}
\usage{
clean_age_population_year(
  data,
  location_col = NULL,
  location = NULL,
  age_col,
  year_col = NULL,
  year = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{location_col}{bare unquoted variable referring to location column}

\item{location}{location name to filter to. If not specified gives all locations.}

\item{age_col}{bare unquoted variable referring to age column}

\item{year_col}{bare unquoted variable referring to year column}

\item{year}{year to filter to. If not specified, gives all years.}
}
\value{
data frame with \code{lower.age.limit} and \code{upper.age.limit} and
optionally filtered down to specific location or year.
}
\description{
Internally used function within \code{\link[=age_population]{age_population()}} to separate age groups
into its lower & upper limit and then filter the data passed to the desired
year and location.
}
\keyword{internal}
