% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-prediction-helpers.R
\name{clean_term_names}
\alias{clean_term_names}
\title{Clean up GAM term names for use in plotting}
\usage{
clean_term_names(term_names)
}
\arguments{
\item{term_names}{character vector of model term names (e.g., 's(offdiag)').}
}
\value{
names like "offdiag", instead of "s(offdiag)"
}
\description{
Clean up GAM term names for use in plotting
}
\examples{
extract_term_names(polymod_setting_models$home) |> clean_term_names()
}
