% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_original_school_work.R
\docType{data}
\name{conmat_original_work_demographics}
\alias{conmat_original_work_demographics}
\title{Original work demographics for conmat}
\format{
A data frame with 121 rows and 2 variables:
\describe{
\item{age}{0 to 120}
\item{work_fraction}{fraction of population working.}
}
}
\source{
{Census of Population and Housing, 2016, TableBuilder}
}
\usage{
conmat_original_work_demographics
}
\description{
An internal dataset containing the original estimates of which fraction of
ages were working in Australia. These can be used inside of
\code{\link[=fit_single_contact_model]{fit_single_contact_model()}} and \code{\link[=fit_setting_contacts]{fit_setting_contacts()}}.
}
\keyword{datasets}
