% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmat-population.R
\name{conmat_population}
\alias{conmat_population}
\title{Define a conmat population}
\usage{
conmat_population(data, age, population)
}
\arguments{
\item{data}{data.frame}

\item{age}{bare name representing the age column}

\item{population}{bare name representing the population column}
}
\value{
a data frame with age and population attributes
}
\description{
A conmat population is a dataframe that stores which columns represent the
age and population information. This is useful as it means we can refer
to this information throughout other functions in the conmat package
without needing to specify or hard code which columns represent the age
and population information.
}
\examples{
perth <- abs_age_lga("Perth (C)")
}
