% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abs_state_work.R
\docType{data}
\name{data_abs_state_work}
\alias{data_abs_state_work}
\title{State wise ABS work population data on different ages for year 2016}
\format{
A data frame with 1044 rows and 6 variables:
\describe{
\item{year}{A number denoting the year as 2016, since the data is based on
2016 Census of Population and Housing.}
\item{state}{a string denoting the abbreviated name of state or territory
name such as 'NSW', 'VIC', 'QLD' and so on.}
\item{age}{a number denoting ages from 0 to 115.}
\item{employed_population}{a number denoting number of people employed
including people with full-time, part-time employment status.}
\item{total_population}{a number depicting the total population belonging
to the age.}
\item{proportion}{a number denoting the measure of the ratio of employed
population and total population belonging to the age i.e,
employed_population/ total_population}
}
}
\source{
{Census of Population and Housing, 2016, TableBuilder}
}
\usage{
data_abs_state_work
}
\description{
A dataset containing Australian Bureau of Statistics labour force population
data by state for 2016. The data sourced from 2016 Census - Employment,
Income and Education through TableBuilder have been randomly adjusted by
the ABS to avoid the release of confidential data.
}
\keyword{datasets}
