% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-davies-age-extended.R
\docType{data}
\name{davies_age_extended}
\alias{davies_age_extended}
\title{Susceptibility and clinical fraction parameters from Davies et al.}
\format{
A data frame of the probability of transmission from a case to a contact. There are 40,804 rows and 6 variables.
\describe{
\item{age}{from 0 to 100}
\item{clinical_fraction}{fraction with clinical symptoms?}
\item{davies_original}{davies original estimates?}
\item{davies_updated}{davies updated estimates?}
}
}
\usage{
davies_age_extended
}
\description{
A dataset containing data from ...(insert url here),
When using this data, ensure that you cite the original authors at
}
\details{
""
}
\keyword{datasets}
