% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_setting_contacts.R
\name{estimate_setting_contacts}
\alias{estimate_setting_contacts}
\title{FUNCTION_TITLE}
\usage{
estimate_setting_contacts(
  contact_data_list,
  survey_population,
  prediction_population = survey_population,
  age_breaks
)
}
\arguments{
\item{contact_data_list}{PARAM_DESCRIPTION}

\item{survey_population}{PARAM_DESCRIPTION}

\item{prediction_population}{PARAM_DESCRIPTION, Default: survey_population}

\item{age_breaks}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
