% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-prediction-helpers.R
\name{extract_term_names}
\alias{extract_term_names}
\title{Helper function to extract term names out of GAM fitted model object.}
\usage{
extract_term_names(fit)
}
\arguments{
\item{fit}{fitted object for one single conmat model setting. E.g., the
home setting.}
}
\value{
character vector of term names
}
\description{
Helper function to extract term names out of GAM fitted model object.
}
\examples{
extract_term_names(polymod_setting_models$home)
}
