% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_polymod.R
\name{extrapolate_polymod}
\alias{extrapolate_polymod}
\title{FUNCTION_TITLE}
\usage{
extrapolate_polymod(population, age_breaks = c(seq(0, 75, by = 5), Inf))
}
\arguments{
\item{population}{PARAM_DESCRIPTION}

\item{age_breaks}{PARAM_DESCRIPTION, Default: c(seq(0, 75, by = 5), Inf)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
