% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_setting_contacts.R
\name{fit_setting_contacts}
\alias{fit_setting_contacts}
\title{Fit a contact model to a survey poulation}
\usage{
fit_setting_contacts(contact_data_list, survey_population)
}
\arguments{
\item{contact_data_list}{contact_data_list}

\item{survey_population}{survey_population}
}
\value{
list of fitted gam models
}
\description{
Fit a contact model to a survey poulation
}
\examples{
# don't run as it takes too long to fit
\dontrun{
contact_model <- fit_setting_contacts(
  contact_data_list = get_polymod_setting_data(), 
  survey_population = get_polymod_population()
)
}
}
\author{
Nicholas Tierney
}
