% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_single_contact_model.R
\name{fit_single_contact_model}
\alias{fit_single_contact_model}
\title{FUNCTION_TITLE}
\usage{
fit_single_contact_model(contact_data, population)
}
\arguments{
\item{contact_data}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[mgcv]{bam}}
\code{\link[stats]{offset}},\code{\link[stats]{c("Poisson", "family")}}
}
