% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tidiers.R
\name{formula-terms}
\alias{formula-terms}
\alias{get_formulas_terms}
\title{Extract out formula terms}
\usage{
get_formulas_terms(model)
}
\arguments{
\item{model}{model object}
}
\description{
Extract out formula terms
}
\examples{
\dontrun{
formula_terms <- get_formulas_terms(sim_m)
formula_terms
}

}
\keyword{internal}
