% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_household_size_population.R
\name{get_abs_household_size_population}
\alias{get_abs_household_size_population}
\title{Get population associated with each household size in an LGA or a state}
\usage{
get_abs_household_size_population(state = NULL, lga = NULL)
}
\arguments{
\item{state}{target Australian state name in abbreviated form, such as "QLD", "NSW", or "TAS"}

\item{lga}{target Australian local government area (LGA) name, such as "Fairfield (C)".  See
\code{\link[=abs_lga_lookup]{abs_lga_lookup()}} for list of lga names}
}
\value{
returns a data frame with household size and the population associated with it in each LGA or state.
}
\description{
Get population associated with each household size in an LGA or a state
}
\examples{
get_abs_household_size_population(state = "NSW")
}
