% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_per_capita_household_size_lga.R
\name{get_abs_per_capita_household_size_lga}
\alias{get_abs_per_capita_household_size_lga}
\title{Get household size distribution based on  LGA name}
\usage{
get_abs_per_capita_household_size_lga(lga = NULL)
}
\arguments{
\item{lga}{target Australian local government area (LGA) name, such as "Fairfield (C)".  See
\code{\link[=abs_lga_lookup]{abs_lga_lookup()}} for list of lga names}
}
\value{
returns a numeric value depicting the per capita household size of the specified LGA
}
\description{
Get household size distribution based on  LGA name
}
\examples{
get_abs_per_capita_household_size_lga(lga = "Fairfield (C)")

}
