% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_population_function.R
\name{get_age_population_function}
\alias{get_age_population_function}
\title{return an interpolating function for populations in 1y age increments}
\usage{
get_age_population_function(population, age_col)
}
\arguments{
\item{population}{population data}

\item{age_col}{age variable}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Return an interpolating function to get populations in 1y age increments
from chunkier distributions produced by socialmixr::wpp_age()
}
\details{
DETAILS
}
\examples{
\dontrun{
get_age_population_function(population, age_col = lower.age.limit)
}
}
