% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_household_size_distribution.R
\name{get_household_size_distribution}
\alias{get_household_size_distribution}
\title{Get household size distribution based on state or LGA name}
\usage{
get_household_size_distribution(state = NULL, lga = NULL)
}
\arguments{
\item{state_name}{target Australian state name in abbreviated form, such as "QLD", "NSW", or "TAS"}

\item{lga_name}{target Australian local government area (LGA) name, such as "Fairfield (C)".  See
\code{\link[=abs_lga_lookup]{abs_lga_lookup()}} for list of lga names}
}
\value{
returns a data frame with household size distributions of a specific state or LGA
}
\description{
Get household size distribution based on state or LGA name
}
\examples{
get_household_size_distribution(lga = "Fairfield (C)")
get_household_size_distribution(state = "NSW")
\dontrun{
# cannot specify both state and LGA
get_household_size_distribution(state = "NSW", lga = "Fairfield (C)")
}
}
