% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_per_capita_household_size.R
\name{get_per_capita_household_size}
\alias{get_per_capita_household_size}
\title{Get per capita household size based on state or LGA name}
\usage{
get_per_capita_household_size(state = NULL, lga = NULL)
}
\arguments{
\item{state}{state name}

\item{lga}{lga name}
}
\value{
Numeric of length 1 - the per capita household size for a given state
or LGA.
}
\description{
Get per capita household size based on state or LGA name
}
\examples{
get_per_capita_household_size(lga = "Fairfield (C)")
get_per_capita_household_size(state = "NSW")
\dontrun{
# cannot specify both state and LGA
get_per_capita_household_size(state = "NSW", lga = "Fairfield (C)")
}
}
\author{
Nick Golding
}
