% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_contact_data.R
\name{get_polymod_contact_data}
\alias{get_polymod_contact_data}
\title{FUNCTION_TITLE}
\usage{
get_polymod_contact_data(
  setting = c("all", "home", "work", "school", "other"),
  ages = 0:100,
  contact_age_imputation = c("sample", "mean", "remove_participant")
)
}
\arguments{
\item{setting}{PARAM_DESCRIPTION, Default: c("all", "home", "work", "school", "other")}

\item{ages}{PARAM_DESCRIPTION, Default: 0:100}

\item{contact_age_imputation}{PARAM_DESCRIPTION, Default: c("sample", "mean", "remove_participant")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{context}},\code{\link[dplyr]{case_when}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{n_distinct}}
\code{\link[stats]{Uniform}}
\code{\link[tidyr]{complete}},\code{\link[tidyr]{expand}}
}
