% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_per_capita_household_size.R
\name{get_polymod_per_capita_household_size}
\alias{get_polymod_per_capita_household_size}
\title{Get polymod per capita household size.}
\usage{
get_polymod_per_capita_household_size()
}
\value{
number, 3.248971
}
\description{
Convenience function to help get the per capita household size.
This is calculated as \code{mean(socialmixr::polymod$participants$hh_size)}.
}
\examples{
get_polymod_per_capita_hosuehold_size()
}
\author{
Nicholas Tierney
}
