% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_population.R
\name{get_polymod_population}
\alias{get_polymod_population}
\title{FUNCTION_TITLE}
\usage{
get_polymod_population()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{context}},\code{\link[dplyr]{mutate-joins}}
\code{\link[socialmixr]{wpp_age}}
\code{\link[stats]{weighted.mean}}
}
