% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_setting_data.R
\name{get_polymod_setting_data}
\alias{get_polymod_setting_data}
\title{FUNCTION_TITLE}
\usage{
get_polymod_setting_data()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
