% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_predictions.R
\name{matrix_to_predictions}
\alias{matrix_to_predictions}
\title{Convert a contact matrix as output by these functions (or socialmixr) into
a long-form tibble}
\usage{
matrix_to_predictions(contact_matrix)
}
\arguments{
\item{contact_matrix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Convert a contact matrix as output by these functions (or socialmixr) into
a long-form tibble
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
