% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_age_matrix}
\alias{new_age_matrix}
\title{Build new age matrix}
\usage{
new_age_matrix(matrix, age_breaks)
}
\arguments{
\item{matrix}{numeric matrix}

\item{age_breaks}{character vector of age breaks, by default the rownames.}
}
\value{
matrix with age breaks attribute
}
\description{
A matrix that knows about its age breaks - which are by default provided as
its rownames. Mostly intended for internal use.
}
\examples{
age_break_names <- c("[0,5)", "[5,10)", "[10, 15)")
age_mat <- matrix(
  runif(9),
  nrow = 3,
  ncol = 3,
  dimnames = list(
    age_break_names,
    age_break_names
  )
)

new_age_matrix(
  age_mat,
  age_breaks = age_break_names
)

}
