% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmat-population.R
\name{new_conmat_population}
\alias{new_conmat_population}
\title{Create a new \code{conmat_population} class object}
\usage{
new_conmat_population(data, age, population)
}
\arguments{
\item{data}{data.frame}

\item{age}{bare column name of numeric data on age}

\item{population}{bare column name of numeric data on population}
}
\value{
object with class \code{conmat_population}
}
\description{
Create a new \code{conmat_population} class object
}
\keyword{internal}
