% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_ngm_setting_matrix}
\alias{new_ngm_setting_matrix}
\title{Establish new BGM setting data}
\usage{
new_ngm_setting_matrix(list_matrix, raw_eigenvalue, scaling, age_breaks)
}
\arguments{
\item{list_matrix}{list of matrices}

\item{raw_eigenvalue}{the raw eigenvalue}

\item{scaling}{scaling factor}

\item{age_breaks}{vector of age breaks}
}
\value{
object with additional (primary) class "ngm_setting_matrix", and attributes for "age_breaks", "scaling", and "raw_eigenvalue".
}
\description{
Establish new BGM setting data
}
