% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_setting_data}
\alias{new_setting_data}
\title{Establish new setting data}
\usage{
new_setting_data(list_df)
}
\arguments{
\item{list_df}{list of data frames}
}
\value{
object with additional (primary) class "setting data" and an "age_breaks attribute.
}
\description{
Establish new setting data
}
