% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-prediction-helpers.R
\name{partial-prediction-sum}
\alias{partial-prediction-sum}
\alias{partial_effects_sum}
\alias{partial_effects_sum.contact_model}
\title{Helper functions to create partial predictive plots for a set of fitted models.}
\usage{
partial_effects_sum(model, ages, ...)

\method{partial_effects_sum}{contact_model}(model, ages, ...)
}
\arguments{
\item{model}{A fitted model, or list of fitted models}

\item{ages}{vector of integer ages}

\item{...}{dots for future extension. Currently not used.}
}
\value{
data frame with 3 columns plus n rows based on expand.grid
combination of ages. The column \code{gam_total_term} is the sum over the coefficients
for that age bracket.
}
\description{
These helper functions exist to make it easier to explore and understand the
impact of each of the covariates used in the conmat GAM model.
}
\details{
Partial predictive plots give a visual representation of the effect of each
covariate on the model, or (equivalently) the effect of each setting on the total
contact matrix. Positive values indicate more contacts in that region of the matrix compared
to the null case, while negative values indicate less.

Scales are not comparable \emph{across} settings, as each setting has it's own intercept term
which is not accounted for in partial effects.
}
\examples{
# Summed up partial effects (y-hat) for a single setting
partials_summed_home <- partial_effects_sum(
    polymod_setting_models$home,
    ages = 1:99
  )
  
autoplot(partials_summed_home)
# summed up partial effects (y-hat) for all settings
partials_summed_setting <- partial_effects_sum(
    polymod_setting_models,
    ages = 1:99
  )
autoplot(partials_summed_setting)
}
