% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setting_matrices.R
\name{plot_setting_matrices}
\alias{plot_setting_matrices}
\title{FUNCTION_TITLE}
\usage{
plot_setting_matrices(
  matrices,
  title = "Setting-specific synthetic contact matrices (all polymod data)"
)
}
\arguments{
\item{matrices}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION, Default: 'Setting-specific synthetic contact matrices (all polymod data)'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{labs}}
\code{\link[patchwork]{plot_layout}},\code{\link[patchwork]{plot_annotation}}
}
