% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymod-data.R
\docType{data}
\name{polymod}
\alias{polymod}
\title{Polymod data}
\format{
An object of class \code{survey} of length 3.
}
\usage{
polymod
}
\description{
Imported from \code{socialmixr}
}
\keyword{datasets}
