% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_polymod_model.R
\docType{data}
\name{polymod_setting_models}
\alias{polymod_setting_models}
\title{Polymod Settings model}
\format{
A model predicting the number of contacts in each of the four
settings, which are: ..., ...
}
\usage{
polymod_setting_models
}
\description{
Add details on how the model was fit using the following code (copy in the
model fitting code from "create-polymod-model.R"):\preformatted{}
}
\details{
\preformatted{}
}
\keyword{datasets}
