% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_contacts.R
\name{predict_contacts}
\alias{predict_contacts}
\title{FUNCTION_TITLE}
\usage{
predict_contacts(model, population, age_breaks = c(seq(0, 75, by = 5), Inf))
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}

\item{age_breaks}{PARAM_DESCRIPTION, Default: c(seq(0, 75, by = 5), Inf)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{arrange}}
}
