% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_contacts_1y.R
\name{predict_contacts_1y}
\alias{predict_contacts_1y}
\title{FUNCTION_TITLE}
\usage{
predict_contacts_1y(model, population, age_min = 0, age_max = 100)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}

\item{age_min}{PARAM_DESCRIPTION, Default: 0}

\item{age_max}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{expand_grid}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
}
