% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_setting_contacts.R
\name{predict_setting_contacts}
\alias{predict_setting_contacts}
\title{Predict setting contacts}
\usage{
predict_setting_contacts(population, contact_model, age_breaks)
}
\arguments{
\item{population}{population}

\item{contact_model}{contact_model}

\item{age_breaks}{age_breaks}
}
\value{
list of matrices
}
\description{
Predict setting contacts
}
\examples{
# don't run as it takes too long to fit
\dontrun{
contact_model <- fit_setting_contacts(
  contact_data_list = get_polymod_setting_data(), 
  survey_population = get_polymod_population()
)

contact_model

contact_model_pred <- predict_setting_contacts(
  population = get_polymod_population(), 
  contact_model = contact_model, 
  age_breaks = c(seq(0, 75, by = 5), Inf)
)
}
}
\author{
Nicholas Tierney
}
