% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_to_matrix.R
\name{predictions_to_matrix}
\alias{predictions_to_matrix}
\title{Predict contacts to matrix}
\usage{
predictions_to_matrix(contact_predictions)
}
\arguments{
\item{contact_predictions}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Predict contacts to matrix
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
