% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-age-population-function-internals.R
\name{prepare_population_for_modelling}
\alias{prepare_population_for_modelling}
\alias{prepare_population_for_modelling.conmat_population}
\alias{prepare_population_for_modelling.data.frame}
\title{Prepare population data for generating an age population function}
\usage{
prepare_population_for_modelling(data, ...)

\method{prepare_population_for_modelling}{conmat_population}(data, ...)

\method{prepare_population_for_modelling}{data.frame}(
  data = data,
  age_col = age_col,
  pop_col = pop_col,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{...}{extra arguments}

\item{age_col}{column of ages}

\item{pop_col}{column of population,}
}
\value{
list of objects, \code{max_bound} \code{pop_model_bounded} \code{bounded_pop} \code{unbounded_pop} for use in \code{\link[=get_age_population_function]{get_age_population_function()}}
}
\description{
Prepares objects for use in \code{\link[=get_age_population_function]{get_age_population_function()}}.
}
\examples{
prepare_population_for_modelling(get_polymod_population())

}
\author{
njtierney
}
\keyword{internal}
