% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{raw_eigenvalue}
\alias{raw_eigenvalue}
\title{Get raw eigvenvalue from NGM matrix}
\usage{
raw_eigenvalue(list_matrix)
}
\arguments{
\item{list_matrix}{object of class \code{ngm_setting_matrix}}
}
\value{
raw eigenvalue
}
\description{
Get raw eigvenvalue from NGM matrix
}
\examples{
# examples not run as they take a long time
\dontrun{
perth <- abs_age_lga("Perth (C)")
perth_contact <- extrapolate_polymod(perth)
perth_ngm <- generate_ngm(
  perth_contact,
  age_breaks = c(seq(0, 85, by = 5), Inf)
)
raw_eigenvalue(perth_ngm)
}
}
