% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{scaling}
\alias{scaling}
\title{Get the scaling from NGM matrix}
\usage{
scaling(list_matrix)
}
\arguments{
\item{list_matrix}{object of class \code{ngm_setting_matrix}}
}
\value{
scaling
}
\description{
This value is \code{scaling <- R_target / R_raw}, where \code{R_target} is the target
R value provided to the NGM, and \code{R_raw} is the raw eigenvalue.
}
\examples{
# examples not run as they take a long time
\donttest{
perth <- abs_age_lga("Perth (C)")
perth_contact <- extrapolate_polymod(perth)
perth_ngm <- generate_ngm(
  perth_contact,
  age_breaks = c(seq(0, 75, by = 5), Inf),
  R_target = 1.5
)
raw_eigenvalue(perth_ngm)
scaling(perth_ngm)
}
}
