% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_age_group}
\alias{separate_age_group}
\title{Separate age groups}
\usage{
separate_age_group(data, age_col)
}
\arguments{
\item{data}{data frame}

\item{age_col}{bare unquoted column referring to age column}
}
\value{
data frame with two extra columns, \code{lower.age.limit} and
\code{upper.age.limit}
}
\description{
An internal function used within \code{\link[=clean_age_population_year]{clean_age_population_year()}} to
separate age groups in a data set into two variables, \code{lower.age.limit},
and \code{upper.age.limit}
}
\keyword{internal}
