% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting-prediction-matrix.R
\name{setting_prediction_matrix}
\alias{setting_prediction_matrix}
\title{Create a setting prediction matrix}
\usage{
setting_prediction_matrix(..., age_breaks)
}
\arguments{
\item{...}{list of matrices}

\item{age_breaks}{age breaks - numeric}
}
\value{
setting prediction matrix
}
\description{
Helper function to create your own setting prediction matrix, which you
may want to use in \code{generate_ngm}, or \code{autoplot}. This class is the
output of functions like \code{extrapolate_polymod}, and
\code{predict_setting_contacts}. We recommend using this function is only for
advanced users, who are creating their own setting prediction matrix.
}
\examples{

age_breaks_0_80_plus <- c(seq(0, 80, by = 10), Inf)
one_by_nine <- matrix(1, nrow = 9, ncol = 9)

x_example <- setting_prediction_matrix(
  home = one_by_nine,
  work = one_by_nine,
  age_breaks = age_breaks_0_80_plus
)

x_example <- setting_prediction_matrix(
  one_by_nine,
  one_by_nine,
  age_breaks = age_breaks_0_80_plus
)

x_example

}
