% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setting_weights.R
\docType{data}
\name{setting_weights}
\alias{setting_weights}
\title{Setting weights computed for transmission probabilities.}
\format{
A named vector of weights relative to home, for home, work, school,
and other
}
\usage{
setting_weights
}
\description{
see ?get_setting_transmission_matrices for details of how to use these
}
\keyword{datasets}
