% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example_vaccination_effect.R
\docType{data}
\name{vaccination_effect_example_data}
\alias{vaccination_effect_example_data}
\title{Example dataset with information on age based vaccination coverage,acquisition and transmission}
\format{
Tibble
}
\usage{
vaccination_effect_example_data
}
\description{
data frame with information on vaccine coverage, efficacy of acquisition/susceptibility and efficacy of transmission/infectiousness
for the ordered age groups from lowest to highest of the next generation matrix
}
\keyword{datasets}
