% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_states.R
\name{abbreviate_states}
\alias{abbreviate_states}
\title{Abbreviate Australian State Names}
\usage{
abbreviate_states(state_names)
}
\arguments{
\item{state_names}{vector of state names in long form}
}
\value{
shortened state names
}
\description{
Given a full name (Title Case) of an Australian State or Territory, produces
the abbreviated state name.
}
\examples{
abbreviate_states("Victoria")
abbreviate_states(c("Victoria", "Queensland"))
}
\seealso{
\code{\link[=unabbreviate_states]{unabbreviate_states()}}
}
