% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_age_lga.R
\name{abs_age_lga}
\alias{abs_age_lga}
\title{Return abs age population data for a given LGA}
\usage{
abs_age_lga(lga_name)
}
\arguments{
\item{lga_name}{lga name - can be a partial match, e.g.,  although the official name might be "Albury (C)", "Albury" is fine. It must also match exactly one LGA. See \code{\link[=check_lga_name]{check_lga_name()}} for more details.}
}
\value{
dataset of columns
}
\description{
Return abs age population data for a given LGA
}
\examples{
abs_age_lga("Albury (C)")
}
