% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abs_state_age.R
\docType{data}
\name{abs_state_age}
\alias{abs_state_age}
\title{ABS state population data for 2020}
\format{
A data frame with 4986 rows and 6 variables:
\describe{
\item{age_group}{age group in five year bins from 0 to 99, then 100+}
\item{state}{state - long state or territory name}
\item{population}{population size}
}
}
\source{
\url{https://www.abs.gov.au/statistics/people/population/national-state-and-territory-population/dec-2020#data-downloads-data-cubes}
}
\usage{
abs_state_age
}
\description{
A dataset containing Australian Bureau of Statistics state level population data for 2020
}
\keyword{datasets}
