% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modelling_features.R
\name{add_modelling_features}
\alias{add_modelling_features}
\title{Add features required for modelling to the dataset}
\usage{
add_modelling_features(contact_data, ...)
}
\arguments{
\item{contact_data}{contact data}

\item{...}{extra dots}
}
\value{
data frame
}
\description{
Add features required for modelling to the dataset
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
