% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rotated_ages.R
\name{add_rotated_ages}
\alias{add_rotated_ages}
\title{apply a 45 degree rotation to age contact pairs}
\usage{
add_rotated_ages(contact_data)
}
\arguments{
\item{contact_data}{contact data}
}
\value{
data set with rotated age contact paris
}
\description{
apply a 45 degree rotation to age contact pairs, to model trends along
the diagonal, and perpendicularly across the diagonal.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
