% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_school_work_participation.R
\name{add_school_work_participation}
\alias{add_school_work_participation}
\title{Add fractions of the population in each age group that attend school/work
(average FTE)}
\usage{
add_school_work_participation(contact_data)
}
\arguments{
\item{contact_data}{contact data}
}
\value{
dataset
}
\description{
Add fractions of the population in each age group that attend school/work
(average FTE) to compute the probability that both participant and
contact attend school/work.
}
\note{
this is fake data that will get replaced with abs data soon
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
