% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_household_contact_matrix.R
\name{adjust_household_contact_matrix}
\alias{adjust_household_contact_matrix}
\title{Adjust Household Contact Matrix

This function is used internally within \code{\link[=predict_setting_contacts]{predict_setting_contacts()}}.
See details below for why we use the per capita adjustment}
\usage{
adjust_household_contact_matrix(
  setting_matrices,
  per_capita_household_size,
  model_percapita_household_size
)
}
\arguments{
\item{setting_matrices}{setting matrix}

\item{per_capita_household_size}{single number of the size of a household}

\item{model_mean_other_householders}{}
}
\value{
contact matrix that has household adjustment
}
\description{
Adjust Household Contact Matrix

This function is used internally within \code{\link[=predict_setting_contacts]{predict_setting_contacts()}}.
See details below for why we use the per capita adjustment
}
\details{
We use Per-capita household size instead of mean household size.
Per-capita household size is different to mean household size, as the
household size averaged over people in the \strong{population}, not over
households, so larger households get upweighted. It is calculated by
taking a distribution of the number of households of each size in a
population, multiplying the size by the household by the household count
to get the number of people with that size of household, and computing
the population-weighted average of household sizes. We use per-capita
household size as it is a more accurate reflection of the average
number of household members a person in the population can have contact
with.
}
\author{
Nick Golding
}
