% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_predicted_contacts.R
\name{aggregate_predicted_contacts}
\alias{aggregate_predicted_contacts}
\title{aggregate predicted contacts from complete 1y resolution}
\usage{
aggregate_predicted_contacts(
  predicted_contacts_1y,
  population,
  age_breaks = c(seq(0, 75, by = 5), Inf)
)
}
\arguments{
\item{predicted_contacts_1y}{contacts in 1 year breaks}

\item{population}{population}

\item{age_breaks}{Default: c(seq(0, 75, by = 5), Inf)}
}
\value{
data frame
}
\description{
aggregate predicted contacts from complete 1y resolution to a stated
resolution must pass in the population to do approppriate weighting of
'from' age groups
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
