% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmat-population.R
\name{as_conmat_population}
\alias{as_conmat_population}
\title{Convert to conmat population}
\usage{
as_conmat_population(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{age}{age column - numeric}

\item{population}{population column - numeric}
}
\description{
Convert to conmat population
}
