% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lga_name.R
\name{check_lga_name}
\alias{check_lga_name}
\title{Check LGA name in Australia}
\usage{
check_lga_name(lga_name, multiple_lga = FALSE)
}
\arguments{
\item{lga_name}{character of length 1}
}
\value{
errors if lga name not in ABS data
}
\description{
Check LGA name in Australia
}
