% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_lga_name}
\alias{check_lga_name}
\title{Check LGA name in Australia}
\usage{
check_lga_name(lga_name, multiple_lga = FALSE)
}
\arguments{
\item{lga_name}{a string denoting the official name of Local Government Area.
For example, 'Albury (C).'}

\item{multiple_lga}{logical response that allows multiple lgas to be checked
if set to \code{TRUE}. Default is FALSE.}
}
\value{
errors if LGA name not in ABS data, otherwise returns nothing
}
\description{
Check LGA name in Australia
}
\examples{
# returns nothing
check_lga_name("Fairfield (C)")
# if you want to check multiple LGAs you must use the `multiple_lga` flag.
check_lga_name(lga_name = c("Fairfield (C)", "Sydney (C)"), 
               multiple_lga = TRUE)
# this will error, so isn't run
\dontrun{
# don't set the `multiple_lga` flag
check_lga_name(lga_name = c("Fairfield (C)", "Sydney (C)"))
# not a fully specified LGA
check_lga_name("Fairfield")
}
}
