% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmat-package.R
\docType{package}
\name{conmat-package}
\alias{conmat}
\alias{conmat-package}
\title{conmat: Builds Contact Matrices using Generalised Additive Models (GAMs) and Population Data}
\description{
Builds contact matrices using Generalised Additive Models (GAMs) and population data, as described in Tierney, Saraswati, Babu, Lydeamore, and Golding (2026) \doi{10.21105/joss.08326}. This package incorporates data that is copyright Commonwealth of Australia (Australian Electoral Commission and Australian Bureau of Statistics) 2020.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://idem-lab.github.io/conmat/}
  \item \url{https://github.com/idem-lab/conmat}
  \item Report bugs at \url{https://github.com/idem-lab/conmat/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nick Golding \email{nick.golding.research@gmail.com} (\href{https://orcid.org/0000-0001-8916-5570}{ORCID})
  \item Aarathy Babu \email{aarathybabu907@gmail.com} (\href{https://orcid.org/0000-0002-6982-5989}{ORCID})
  \item Chitra Saraswati \email{chitra.saraswati@telethonkids.org.au} (\href{https://orcid.org/0000-0002-8159-0414}{ORCID})
  \item Michael Lydeamore \email{michael.lydeamore@monash.edu} (\href{https://orcid.org/0000-0001-6515-827X}{ORCID})
}

Other contributors:
\itemize{
  \item Commonwealth of Australia AEC [copyright holder]
  \item Australian Bureau of Statistics ABS [copyright holder]
}

}
\keyword{internal}
