% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_setting_contacts.R
\name{estimate_setting_contacts}
\alias{estimate_setting_contacts}
\title{Get predicted contact matrices for each setting and combined}
\usage{
estimate_setting_contacts(
  contact_data_list,
  survey_population,
  prediction_population = survey_population,
  age_breaks,
  per_capita_household_size = NULL
)
}
\arguments{
\item{contact_data_list}{PARAM_DESCRIPTION}

\item{survey_population}{PARAM_DESCRIPTION}

\item{prediction_population}{PARAM_DESCRIPTION, Default: survey_population}

\item{age_breaks}{PARAM_DESCRIPTION}

\item{per_capita_household_size}{PARAM_DESCRIPTION, Default : NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Given a named list of contact datasets (with names giving the setting, and
assumed to together make up the full set of contacts for individuals in the
survey), a representative population distribution for the survey, and a set
of age breaks at which to aggregate contacts, return a set of predicted
contact matrices for each setting, and for all combined.
}
