% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_polymod.R
\name{extrapolate_polymod}
\alias{extrapolate_polymod}
\title{Fit all-of-polymod model and extrapolate to a given population an age breaks}
\usage{
extrapolate_polymod(
  population,
  age_breaks = c(seq(0, 75, by = 5), Inf),
  per_capita_household_size = NULL
)
}
\arguments{
\item{population}{a \code{conmat_population} object, specifying the \code{age}
and \code{population} characteristics. Or a data frame with \code{lower.age.limit}
and \code{population} columns. See \code{get_polymod_population()} for an example
of this data.}

\item{age_breaks}{vector depicting age values. Default value is
\code{c(seq(0, 75, by = 5), Inf)}}

\item{per_capita_household_size}{Optional (defaults to NULL). When set, it
adjusts the household contact matrix by some per capita household size.
To set it, provide a single number, the per capita household size. More
information is provided below in Details. See
\code{\link[=get_abs_per_capita_household_size]{get_abs_per_capita_household_size()}} function for a helper for Australian
data with a workflow on how to get this number.}
}
\value{
Returns setting-specific and combined contact matrices for the
desired ages.
}
\description{
Uses \code{\link[=estimate_setting_contacts]{estimate_setting_contacts()}} to fit a contact model on the data from
polymod and later extrapolate on to a desired population. Note that this
function is parallelisable with \code{future}, and will be impacted by any
\code{future} plans provided.
}
\details{
Also note that since this model uses the already fit \code{polymod_setting_models}
data, which has been fit using symmetrical model terms, if you want to
fit a model with asymmetric model terms, you will need to go through
the full process of building new models. You can find this detail in last
section of the vignette "example pipeline".
}
\examples{
\donttest{
polymod_population <- get_polymod_population()
synthetic_settings_5y_polymod <- extrapolate_polymod(
  population = polymod_population
)
synthetic_settings_5y_polymod
synthetic_settings_5y_fairfield <- extrapolate_polymod(
  population = abs_age_lga("Fairfield (C)")
)
synthetic_settings_5y_fairfield
}
}
