% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_single_contact_model.R
\name{fit_single_contact_model}
\alias{fit_single_contact_model}
\title{fit a single GAM contact model to a dataset}
\usage{
fit_single_contact_model(contact_data, population)
}
\arguments{
\item{contact_data}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
fit a single GAM contact model to a dataset
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
