% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-ngm.R
\name{generate_ngm}
\alias{generate_ngm}
\title{Get next generation contact matrices}
\usage{
generate_ngm(state_name = NULL, lga_name = NULL, age_breaks, R_target)
}
\arguments{
\item{state_name}{target state name}

\item{lga_name}{target LGA name}

\item{age_breaks}{age breaks}

\item{R_target}{target reproduction number}
}
\description{
Get next generation contact matrices
}
\examples{
\dontrun{
ngm_NSW <- generate_ngm(
  state_name = "NSW",
  age_breaks = c(seq(0, 85, by = 5), Inf),
  R_target = 1.5
)
ngm_fairfield <- generate_ngm(
  lga_name = "Fairfield (C)",
  age_breaks = c(seq(0, 85, by = 5), Inf),
  R_target = 1.5
)
}
}
