% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_abs_age_work.R
\name{get_data_abs_age_work}
\alias{get_data_abs_age_work}
\title{Return  data on employed population for a given age and state or lga of Australia}
\usage{
get_data_abs_age_work(age, state = NULL, lga = NULL)
}
\arguments{
\item{age}{a numeric or numeric vector denoting ages between 0 to 115}

\item{state}{target Australian state name in abbreviated form, such as "QLD", "NSW", or "TAS"}

\item{lga}{target Australian local government area (LGA) name, such as "Fairfield (C)".  See
\code{\link[=abs_lga_lookup]{abs_lga_lookup()}} for list of lga names}
}
\value{
data set with information on the number of employed people belonging to a particular age, its total population
and the corresponding proportion.
}
\description{
Return  data on employed population for a given age and state or lga of Australia
}
\examples{
get_data_abs_age_work(state="NSW",age=10:20)
get_data_abs_age_work(lga="Barcoo (S)",age=39)
}
