% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_contact_data.R
\name{get_polymod_contact_data}
\alias{get_polymod_contact_data}
\title{Format polymod data and filter contacts to certain settings}
\usage{
get_polymod_contact_data(
  setting = c("all", "home", "work", "school", "other"),
  countries = c("Belgium", "Finland", "Germany", "Italy", "Luxembourg", "Netherlands",
    "Poland", "United Kingdom"),
  ages = 0:100,
  contact_age_imputation = c("sample", "mean", "remove_participant")
)
}
\arguments{
\item{setting}{Default: c("all", "home", "work", "school", "other")}

\item{countries}{countries to extract data from}

\item{ages}{Default: 0:100}

\item{contact_age_imputation}{Default: c("sample", "mean", "remove_participant")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Format polymod data and filter contacts to certain settings
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
