% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_contact_data.R
\name{get_polymod_contact_data}
\alias{get_polymod_contact_data}
\title{Format polymod data and filter contacts to certain settings}
\usage{
get_polymod_contact_data(
  setting = c("all", "home", "work", "school", "other"),
  ages = 0:100,
  contact_age_imputation = c("sample", "mean", "remove_participant")
)
}
\arguments{
\item{setting}{PARAM_DESCRIPTION, Default: c("all", "home", "work", "school", "other")}

\item{ages}{PARAM_DESCRIPTION, Default: 0:100}

\item{contact_age_imputation}{PARAM_DESCRIPTION, Default: c("sample", "mean", "remove_participant")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Format polymod data and filter contacts to certain settings
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
