% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_population.R
\name{get_polymod_population}
\alias{get_polymod_population}
\title{return the polymod-average population age distribution in 5y}
\usage{
get_polymod_population()
}
\value{
data frame
}
\description{
return the polymod-average population age distribution in 5y increments
(weight country population distributions by number of participants)
note that we don't want to weight by survey age distributions for this, since
the total number of \emph{participants} represents the sampling
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
