% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_population.R
\name{get_polymod_population}
\alias{get_polymod_population}
\title{Return the polymod-average population age distribution in 5y}
\usage{
get_polymod_population(
  countries = c("Belgium", "Finland", "Germany", "Italy", "Luxembourg", "Netherlands",
    "Poland", "United Kingdom")
)
}
\arguments{
\item{countries}{countries to extract data from. Default is to get: Belgium,
Finland, Germany, Italy, Luxembourg, Netherlands, Poland, and
United Kingdom.}
}
\value{
data frame with two columns: \code{lower.age.limit} and \code{population}
}
\description{
returns the polymod-average population age distribution in
5y increments (weight country population distributions by number of
participants). Note that we don't want to weight by survey age
distributions for this, since the total number of \emph{participants}
represents the sampling. It uses the participant data from the polymod
survey as well as the age specific population data from \code{socialmixr}
R package to return the age specific average population of different,
countries weighted by the number of participants from those countries who
participated in the polymod survey.
}
\examples{
get_polymod_population()
get_polymod_population("Belgium") 
get_polymod_population("United Kingdom") 
get_polymod_population("Italy") 
}
