% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_setting_data.R
\name{get_polymod_setting_data}
\alias{get_polymod_setting_data}
\title{Get polymod setting data}
\usage{
get_polymod_setting_data(
  countries = c("Belgium", "Finland", "Germany", "Italy", "Luxembourg", "Netherlands",
    "Poland", "United Kingdom")
)
}
\arguments{
\item{countries}{countries to extract data from}
}
\value{
A list of data frames, of the polymod data. One list per setting:
"home", "work", "school", and "other".
}
\description{
\code{get_polymod_setting_data()} acts as an extension of
\code{get_polymod_contact_data()}, and extracts the setting wise contact data
on the desired country, as a list.
}
\examples{
get_polymod_setting_data()
get_polymod_setting_data("Belgium")
}
