% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polymod_setting_data.R
\name{get_polymod_setting_data}
\alias{get_polymod_setting_data}
\title{Get polymod setting data}
\usage{
get_polymod_setting_data()
}
\value{
polymod data
}
\description{
Get polymod setting data
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
