% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_predictions.R
\name{matrix_to_predictions}
\alias{matrix_to_predictions}
\title{FUNCTION_TITLE}
\usage{
matrix_to_predictions(contact_matrix)
}
\arguments{
\item{contact_matrix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{as_tibble}}
\code{\link[tidyr]{pivot_longer}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{across}},\code{\link[dplyr]{reexports}}
}
