% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-prediction-helpers.R
\name{partial-prediction}
\alias{partial-prediction}
\alias{partial_effects}
\alias{partial_effects.contact_model}
\alias{partial_effects.setting_contact_model}
\alias{partial_effects_sum.setting_contact_model}
\title{Helper functions to create partial predictive plots.}
\usage{
partial_effects(model, ages, ...)

\method{partial_effects}{contact_model}(model, ages, ...)

\method{partial_effects}{setting_contact_model}(model, ages, ...)

\method{partial_effects_sum}{setting_contact_model}(model, ages, ...)
}
\arguments{
\item{model}{A fitted model, or list of fitted models.}

\item{ages}{vector of integer ages.}

\item{...}{extra arguments. Currently not used.}
}
\value{
data frame with 20 columns plus n rows based on expand.grid
combination of ages. Contains transformed coefficients from ages.
}
\description{
These helper functions exist to make it easier to explore and understand the
impact of each of the covariates used in the conmat GAM model.
}
\details{
Partial predictive plots give a visual representation of the effect of each
covariate on the model, or (equivalently) the effect of each setting on
the total contact matrix. Positive values indicate more contacts in that
region of the matrix compared to the null case, while negative values
indicate less.

Scales are not comparable \emph{across} settings, as each setting has it's own
intercept term which is not accounted for in partial effects.
}
\examples{
partials_home <- partial_effects(
  polymod_setting_models$home, 
  ages = 1:99
  )
autoplot(partials_home)
# partial effects for all settings
partials_setting <- partial_effects(
  polymod_setting_models, 
  ages = 1:99
  )
autoplot(partials_setting)
}
