% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Use ggplot to plot a matrix in the output format}
\usage{
plot_matrix(matrix)
}
\arguments{
\item{matrix}{matrix}
}
\value{
a ggplot object
}
\description{
Use ggplot to plot a matrix in the output format
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
