% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setting_matrices.R
\name{plot_setting_matrices}
\alias{plot_setting_matrices}
\title{Plot setting matrices using ggplot2}
\usage{
plot_setting_matrices(
  matrices,
  title = "Setting-specific synthetic contact matrices"
)
}
\arguments{
\item{matrices}{matrix}

\item{title}{Title to give to plot setting matrices. Default value is: Setting-specific synthetic contact matrices (all polymod data)'}
}
\value{
a ggplot
}
\description{
Plot setting matrices using ggplot2
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
