% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_polymod_model.R
\docType{data}
\name{polymod_setting_models}
\alias{polymod_setting_models}
\title{Polymod Settings models}
\format{
An object of class \code{setting_contact_model} (inherits from \code{list}) of length 4.
}
\usage{
polymod_setting_models
}
\description{
A data object containing a list of fitted gam models
predicting the number of contacts in each of the four settings which are
"home","work","school" and "other". For more details on model fitting,
see \code{\link[=fit_setting_contacts]{fit_setting_contacts()}}. This object has been provided as data to
avoid recomputing a relatively common type of model for use with \code{conmat}.
}
\examples{
\donttest{
# code used to produce this data
library(conmat)
set.seed(2025 - 11 - 28)
polymod_contact_data <- get_polymod_setting_data()
polymod_survey_data <- get_polymod_population()
polymod_setting_models <- fit_setting_contacts(
  contact_data_list = polymod_contact_data,
  population = polymod_survey_data
)
}

}
\seealso{
\code{\link[=fit_setting_contacts]{fit_setting_contacts()}}
}
\keyword{datasets}
