% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_contacts.R
\name{predict_contacts}
\alias{predict_contacts}
\title{predict number of contacts in all combinations of the age groups specified
by age_breaks, given a model and a population distribution of the time/place
to predict to}
\usage{
predict_contacts(model, population, age_breaks = c(seq(0, 75, by = 5), Inf))
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}

\item{age_breaks}{PARAM_DESCRIPTION, Default: c(seq(0, 75, by = 5), Inf)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
predict number of contacts in all combinations of the age groups specified
by age_breaks, given a model and a population distribution of the time/place
to predict to
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
