% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_contacts_1y.R
\name{predict_contacts_1y}
\alias{predict_contacts_1y}
\title{predict contacts to a given population at full 1y resolution}
\usage{
predict_contacts_1y(model, population, age_min = 0, age_max = 100)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}

\item{age_min}{PARAM_DESCRIPTION, Default: 0}

\item{age_max}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
predict contacts to a given population at full 1y resolution
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
