% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_setting_contacts.R
\name{predict_setting_contacts}
\alias{predict_setting_contacts}
\title{Predict setting contacts}
\usage{
predict_setting_contacts(
  population,
  contact_model,
  age_breaks,
  per_capita_household_size = NULL,
  ...
)
}
\arguments{
\item{population}{population}

\item{contact_model}{contact_model}

\item{age_breaks}{age_breaks}

\item{per_capita_household_size}{Optional (defaults to NULL). When set, it
adjusts the household contact matrix by some per capital household size.
To set it, provide a single number, which is the per capita household size.
See \code{get_per_capita_household_size()} function for a helper for Australian
data with a workflow on how to ge this number.}

\item{...}{extra arguments to pass to the \code{per_capita_household_size}
adjustment - see \code{adjust_household_contact_matrix} for details.
Default values for this are \code{household_contact_rate = 1} and
\code{model_mean_other_householders = 2.248971}.}
}
\value{
List of setting matrices
}
\description{
Predict setting contacts
}
\examples{
# don't run as it takes too long to fit
\dontrun{
fairfield_age_pop <- abs_age_lga("Fairfield (C)")
fairfield_age_pop

polymod_contact_data <- get_polymod_setting_data()
polymod_survey_data <- get_polymod_population()

setting_models <- fit_setting_contacts(
  contact_data_list = polymod_contact_data,
  population = polymod_survey_data
)

synthetic_settings_5y_fairfield <- predict_setting_contacts(
  population = fairfield_age_pop,
  contact_model = setting_models,
  age_breaks = c(seq(0, 85, by = 5), Inf)
)

fairfield_hh_size <- get_per_capita_household_size(lga = "Fairfield (C)")
fairfield_hh_size

synthetic_settings_5y_fairfield_hh <- predict_setting_contacts(
  population = fairfield_age_pop,
  contact_model = setting_models,
  age_breaks = c(seq(0, 85, by = 5), Inf),
  per_capita_household_size = fairfield_hh_size
)
}
}
\author{
Nicholas Tierney
}
