% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_setting_contacts.R
\name{predict_setting_contacts}
\alias{predict_setting_contacts}
\title{Predict setting contacts}
\usage{
predict_setting_contacts(
  population,
  contact_model,
  age_breaks,
  per_capita_household_size = NULL
)
}
\arguments{
\item{population}{population}

\item{contact_model}{contact_model}

\item{age_breaks}{age_breaks}

\item{per_capita_household_size}{single number, which is the per capita household size. See \code{get_per_capita_household_size()} function for a helper for Australian data.}
}
\value{
list of matrices
}
\description{
Predict setting contacts
}
\examples{
# don't run as it takes too long to fit
\dontrun{
contact_model <- fit_setting_contacts(
  contact_data_list = get_polymod_setting_data(), 
  survey_population = get_polymod_population()
)

contact_model

contact_model_pred <- predict_setting_contacts(
  population = get_polymod_population(), 
  contact_model = contact_model, 
  age_breaks = c(seq(0, 75, by = 5), Inf)
)

# Adjust the household size distribution
synthetic_settings_5y_fairfield <- predict_setting_contacts(
  population = fairfield_age_pop,
  contact_model = setting_models,
  age_breaks = c(seq(0, 85, by = 5), Inf),
  household_size_distribution = fairfield_household_sizes
)
}
}
\author{
Nicholas Tierney
}
