% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_to_matrix.R
\name{predictions_to_matrix}
\alias{predictions_to_matrix}
\title{FUNCTION_TITLE}
\usage{
predictions_to_matrix(contact_predictions)
}
\arguments{
\item{contact_predictions}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{pivot_wider}}
\code{\link[tibble]{rownames}}
}
