% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prem_contact_matrices.R
\docType{data}
\name{prem_germany_contact_matrices}
\alias{prem_germany_contact_matrices}
\title{Contact matrices as calculated by Prem. et al.}
\format{
A list with 5 elements:
\describe{
\item{home}{A 16x16 matrix containing the number of home contacts, by 5
year age group}
\item{work}{A 16x16 matrix containing the number of workplace contacts, by
5 year age group}
\item{school}{A 16x16 matrix containing the number of school contacts, by 5
year age group}
\item{other}{A 16x16 matrix containing the number of other contacts, by 5
year age group}
}
All age groups are 5 year age bands, from 0 to 80.
}
\source{
\url{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005697}
}
\usage{
prem_germany_contact_matrices
}
\description{
Contact matrices as calculated by Prem. et al. PLoS Computational Biology.
DOI: 10.1371/journal.pcbi.1005697
}
\keyword{datasets}
