% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{raw_eigenvalue}
\alias{raw_eigenvalue}
\title{Get raw eigvenvalue from NGM matrix}
\usage{
raw_eigenvalue(list_matrix)
}
\arguments{
\item{list_matrix}{object of class \code{ngm_setting_matrix}}
}
\value{
raw eigenvalue
}
\description{
Get raw eigvenvalue from NGM matrix
}
\examples{
perth <- abs_age_lga("Perth (C)")
perth_contact <- extrapolate_polymod(perth)
perth_contact
# must be the same age breaks as in perth_contact - 0-75
perth_ngm <- generate_ngm(
  perth_contact,
  age_breaks = c(seq(0, 75, by = 5), Inf),
  R_target = 1.5
)
raw_eigenvalue(perth_ngm)
}
