% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setting_weights.R
\docType{data}
\name{setting_weights}
\alias{setting_weights}
\title{Setting weights computed for transmission probabilities from Eyre et al.}
\format{
A named vector of weights relative to home, for home, work, school,
and other
}
\usage{
setting_weights
}
\description{
A vector that contains weights to accompany the digitised transmission
probabilities from "The impact of SARS-CoV-2 vaccination on Alpha & Delta
variant transmission", by David W Eyre, Donald Taylor, Mark Purver, David
Chapman, Tom Fowler, Koen B Pouwels, A Sarah Walker, Tim EA Peto
(\doi{https://doi.org/10.1101/2021.09.28.21264260}.
}
\details{
see ?get_setting_transmission_matrices for details of how to use these
}
\keyword{datasets}
