% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unabbreviate_states.R
\name{unabbreviate_states}
\alias{unabbreviate_states}
\title{Un-bbreviate Australian state names}
\usage{
unabbreviate_states(state_names)
}
\arguments{
\item{state_names}{vector of state names in short form}
}
\value{
Longer state names
}
\description{
Un-bbreviate Australian state names
}
\examples{
abbreviate_states("VIC")
}
