% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example_vaccination_effect.R
\docType{data}
\name{vaccination_effect_example_data}
\alias{vaccination_effect_example_data}
\title{Example dataset with information on age based vaccination coverage,
acquisition and transmission}
\format{
A data frame with 17 rows and 4 variables
\describe{
\item{age_band}{character. age bands: 0-4,5-11, 12-15, 16-19, 20-24, etc}
\item{coverage}{example vaccination coverage, between 0-1}
\item{acquisition}{example acquisition coverage, between 0-1}
\item{transmission}{example transmission coverage, between 0-1}
}
}
\usage{
vaccination_effect_example_data
}
\description{
data frame with information on vaccine coverage, efficacy of
acquisition/susceptibility and efficacy of transmission/infectiousness
for the ordered age groups from lowest to highest of the next generation
matrix.
}
\keyword{datasets}
